restructuring of testimonial component
