+++ src/components/slider/index.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
 import Slider from "react-slick";
+import { Link } from 'gatsby'
 // Styles Slider
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 
-import BoxTestimonial from '../box'
 import {
   Item,
   ContentSlider
@@ -13,7 +13,7 @@ import {
 export default class SliderTestimonial extends Component {
   render() {
     const settings = {
-      dots: true,
+      autoplay: true,
       infinite: true,
       speed: 500,
       slidesToShow: 3,
@@ -48,23 +48,37 @@ export default class SliderTestimonial extends Component {
     return (
       <ContentSlider>
         <Slider {...settings}>
+
           <Item>
-            <BoxTestimonial />
+            <Link to="/">
+                <h1>1</h1>
+            </Link>
+          </Item>
+
+          <Item>
+            <Link to="/">
+                <h1>2</h1>
+            </Link>
           </Item>
           <Item>
-            <BoxTestimonial />
+            <Link to="/">
+                <h1>3</h1>
+            </Link>
           </Item>
           <Item>
-            <BoxTestimonial />
+            <Link to="/">
+                <h1>4</h1>
+            </Link>
           </Item>
           <Item>
-            <BoxTestimonial />
+            <Link to="/">
+                <h1>5</h1>
+            </Link>
           </Item>
           <Item>
-            <BoxTestimonial />
-          </Item>
-          <Item>
-            <BoxTestimonial />
+            <Link to="/">
+                <h1>6</h1>
+            </Link>
           </Item>
         </Slider>
       </ContentSlider>
