--- src/components/testimonial/box/index.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-import { Box } from './styledComponents'
-
-const BoxTestimonial = ({ data }) => (
-  <Box>
-    box testimonial
-  </Box>
-)
-
-
-export default BoxTestimonial
