--- src/components/testimonial/box/styledComponents.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-// import breakpoint from 'styled-components-breakpoint';
-
-export const Box = styled.div`
-  background: #ccc;
-  width: 85%;
-  margin: 0 auto;
-  border-radius: 10px;
-  padding: 10px;
-`;
