+++ src/components/testimonial/styledComponents.js
@@ -1,8 +1,10 @@
 import styled from 'styled-components';
 // import breakpoint from 'styled-components-breakpoint';
-// import { APP_COLORS } from '../utils/constants';
 
-export const Container = styled.div`
-  max-width: 960px;
+export const Box = styled.div`
+  background: #ccc;
+  width: 85%;
   margin: 0 auto;
+  border-radius: 10px;
+  padding: 10px;
 `;
