+++ src/components/title.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { H1, H2 } from './styledComponents'
 
-const Title = ({ type, text }) => {
+const Title = ({ type, text, align }) => {
     switch(type) {
         case 2 : return (<H2>{text}</H2>);
         default: return (<H1>{text}</H1>);
