+++ src/data/home.json
@@ -1,4 +1,21 @@
 {
-    "title": "Sancrisoft",
-    "slogan": "Developing Websites"
+  "title": "Sancrisoft",
+  "slogan": "Developing Websites",
+  "testimonials": [
+    {
+      "description":"Lorem ipsum dolor sit amet, consectetur adipisicing elit, aliqua.Ut enim ad minimo cborum.",
+      "name":"juan perez",
+      "date": "12/04/2017"
+    },
+    {
+      "description":"Lorem ipsum dolor sit amet, consectetur adipisiciminimo cborum.",
+      "name":"pedro pablo",
+      "date": "24/12/2016"
+    },
+    {
+      "description":"Lorem ipsum dolor sit ar adipisicing elit, aliqua.Ut enim ad minimo cborum.",
+      "name":"camilo ortiz",
+      "date": "03/10/2018"
+    }
+  ]
 }
