+++ src/pages/styledComponents.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+// import breakpoint from 'styled-components-breakpoint';
+
+export const ContTestimonial = styled.section`
+  display: block;
+`;
+export const Container = styled.div`
+  margin: 0 auto;
+  max-width: 960px;
+`;
